requirements.txt
asgiref==3.11.0
channels==4.3.2
click==8.3.1
Django==6.0.1
djangorestframework==3.16.1
h11==0.16.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
uvicorn==0.40.0
djangorestframework-simplejwt==5.5.1